GOB: move video code into seperate folder for video
move all video related source code files into an folder for better code overview, like for save or sound
