+++ engines/gob/cmpfile.cpp
@@ -32,7 +32,7 @@
 #include "gob/gob.h"
 #include "gob/util.h"
 #include "gob/surface.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/dataio.h"
 #include "gob/rxyfile.h"
 #include "gob/cmpfile.h"
