+++ engines/gob/demos/demoplayer.cpp
@@ -35,7 +35,7 @@
 #include "gob/util.h"
 #include "gob/draw.h"
 #include "gob/inter.h"
-#include "gob/videoplayer.h"
+#include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
 
 namespace Gob {
