+++ engines/gob/init.h
@@ -29,7 +29,7 @@
 #define GOB_INIT_H
 
 #include "gob/gob.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 namespace Gob {
