+++ engines/gob/init_geisha.cpp
@@ -31,7 +31,7 @@
 #include "gob/init.h"
 #include "gob/global.h"
 #include "gob/draw.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 namespace Gob {
