+++ engines/gob/inter_fascin.cpp
@@ -38,8 +38,8 @@
 #include "gob/expression.h"
 #include "gob/script.h"
 #include "gob/palanim.h"
-#include "gob/video.h"
-#include "gob/videoplayer.h"
+#include "gob/video/video.h"
+#include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
 
 namespace Gob {
