+++ engines/gob/inter_v1.cpp
@@ -44,8 +44,8 @@
 #include "gob/mult.h"
 #include "gob/palanim.h"
 #include "gob/scenery.h"
-#include "gob/video.h"
-#include "gob/videoplayer.h"
+#include "gob/video/video.h"
+#include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
 
 namespace Gob {
