+++ engines/gob/minigames/geisha/penetration.cpp
@@ -31,7 +31,7 @@
 #include "gob/util.h"
 #include "gob/palanim.h"
 #include "gob/draw.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/decfile.h"
 #include "gob/cmpfile.h"
 #include "gob/anifile.h"
