+++ engines/gob/mult.cpp
@@ -36,8 +36,8 @@
 #include "gob/script.h"
 #include "gob/palanim.h"
 #include "gob/scenery.h"
-#include "gob/video.h"
-#include "gob/videoplayer.h"
+#include "gob/video/video.h"
+#include "gob/video/videoplayer.h"
 #include "gob/inter.h"
 #include "gob/sound/sound.h"
