+++ engines/gob/palanim.cpp
@@ -30,7 +30,7 @@
 #include "gob/palanim.h"
 #include "gob/global.h"
 #include "gob/util.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 namespace Gob {
