+++ engines/gob/pregob/onceupon/onceupon.cpp
@@ -31,7 +31,7 @@
 #include "gob/dataio.h"
 #include "gob/surface.h"
 #include "gob/draw.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/anifile.h"
 #include "gob/aniobject.h"
