+++ engines/gob/pregob/onceupon/parents.cpp
@@ -30,7 +30,7 @@
 #include "gob/dataio.h"
 #include "gob/palanim.h"
 #include "gob/draw.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 #include "gob/sound/sound.h"
