+++ engines/gob/pregob/onceupon/stork.cpp
@@ -30,7 +30,7 @@
 #include "gob/gob.h"
 #include "gob/surface.h"
 #include "gob/anifile.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 #include "gob/pregob/onceupon/stork.h"
