+++ engines/gob/save/savefile.cpp
@@ -33,7 +33,7 @@
 
 #include "gob/gob.h"
 #include "gob/save/savefile.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/inter.h"
 #include "gob/variables.h"
