+++ engines/gob/save/savehandler.cpp
@@ -32,7 +32,7 @@
 #include "gob/save/savefile.h"
 #include "gob/save/saveconverter.h"
 #include "gob/global.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/draw.h"
 #include "gob/variables.h"
 #include "gob/inter.h"
