+++ engines/gob/scenery.cpp
@@ -37,7 +37,7 @@
 #include "gob/resources.h"
 #include "gob/inter.h"
 #include "gob/map.h"
-#include "gob/videoplayer.h"
+#include "gob/video/videoplayer.h"
 
 namespace Gob {
