+++ engines/gob/util.cpp
@@ -35,8 +35,8 @@
 #include "gob/dataio.h"
 #include "gob/draw.h"
 #include "gob/game.h"
-#include "gob/video.h"
-#include "gob/videoplayer.h"
+#include "gob/video/video.h"
+#include "gob/video/videoplayer.h"
 #include "gob/sound/sound.h"
 
 namespace Gob {
