+++ engines/gob/video/video.cpp
@@ -34,7 +34,7 @@
 #include "graphics/surface.h"
 
 #include "gob/gob.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 #include "gob/global.h"
 #include "gob/util.h"
 #include "gob/dataio.h"
