+++ engines/gob/video/video_v2.cpp
@@ -28,7 +28,7 @@
 #include "common/endian.h"
 
 #include "gob/gob.h"
-#include "gob/video.h"
+#include "gob/video/video.h"
 
 namespace Gob {
